requirements.txt
fastapi==0.110.0
python-weather==2.0.3
kafka-python==2.0.2
uvicorn==0.30.1

# Nested deps
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
click==8.1.7
frozenlist==1.4.1
h11==0.14.0
idna==3.7
multidict==6.0.5
pip==24.1
setuptools==70.1.0
typing-extensions==4.12.2
yarl==1.9.4
